record decorator became multi-functional
